util: Add MurderDeathKill function
Part of ensuring that everything shuts down cleanly is also ensuring that
we do not leave any child processes running within the root. This will help
in unmounting the overlayfs when a build bails, or in future when we trap
ctrl+c to nuke all the children and tear down properly.

This also sounded far more violent than I intended.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>