builder: Start working on a Manager implementation
At the moment the code is kind of all over the place and highly
duplicated. We're also still unable to cleanly tear down our session
when ctrl+c'd due to the pissed spaghetti nature of the codebase.

We'll move towards a builder.Manager instance which can be responsible
for keeping everything, well. Managed. :P

Signed-off-by: Ikey Doherty <ikey@solus-project.com>