build: Set ownership on the collected files to the sudo user
In most cases solbuild will be run via sudo, so the resulting files are
owned by root. This isn't particulary helpful for the user when they don't
have any write permissions on the new files.

To make life easier, when we encounter SUDO_GID SUDO_UID variables in the
environment, we'll chown the collected files to that user.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>