build: Add support for ypkg reproducible builds
ypkg will now take a -t parameter with a unix timestamp, which for
solbuild will be the time of the last git tag in the repository.
This means that on two different machines building the same tag,
the hash should be identical for both packages.

This should also help with deltas too because less files change.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>