manager: Fix broken SetTmpfs call
The build/index methods take their tmpfs knowledge directly from the
config instance, which will default to tmpfs being disabled. The -t
option is provided at runtime to allow users to enable tmpfs on
demand.

This change ensures we update the config which is latered used to set
the tmpfs mode in the overlay.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>