Remove pending items, we found resolutions
The builder will not use tmpfs to build as we can't introspect it
after, so there's little sense in a tmpfs option. As for the debuginfo
situation, it's only an issue with the legacy format. For now a workaround
has been identified for our kernel package, and there is very little point
in investing any energy into legacy fixes, given that sol will be here
soon to deprecate the old format entirely.

Ypkg rules.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>