source/simple: Finish implementing FTP support
We utilize the existing URL object we parsed earlier to find out any
relevant FTP credentials. If they're not supplied, we'll default to
logging in as anonymous/anonymous. Once that's done we use the vendored
jlaffaye/ftp library to connect to the FTP server, list the path to get
it's size, and then RETR the file to the local filesystem.

We utilize much of the same framework and method as the existing HTTP
fetcher so that even FTP downloads use the progressbar system.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>