source/simple: Use CURL library for downloading non ftp
The current implemention in native golang hits some pretty serious errors
when going up against the dreaded sourceforge. Go's library tries to be
as forgiving and as standard as possible, but just cannot download normal
URLs from sourceforge, it instead downloads HTML pages.

This is entirely sourceforge's fault, of course, but we'll use libcurl
to handle the heavy lifting as it's very fault tolerant in the face of
extreme derp.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>