cmd: Add a new delete-cache command
To resolve #3, we add a new delete-cache command to delete all of the
build roots in /var/cache/solbuild. We also provide the "-a", "--all"
flag which will additionally remove the package, ccache, and source
trees whilst leaving the builder images intact.

This ensures we have a safe method in which to clean up the roots.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>