builder/history: Do not rely on lexicographical sort for history
We must instead rely on sorting the numeric releases in the history's
parsed history for every tag. While this potentially adds time to the
initial build (ms) it ensures we end up with a neatly curated list of
sorted-by-release update entries, which can then be truncated to the
maximum permitted changelog length.

This fixes a bug in some Solus packages whereby the head entry is actually
missing in the changelog due to invalid sorting, so a fake changelog entry
is being generated instead.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>