history: Improve timestamps by changing them less often
Currently we change the version timestamp on every single bump of the
package, which causes a lot of unnecessary noise. Further, this impacts
the usefuless of delta packages by skewing their timestamps from simple
bump-rebuilds.

We now set the timestamp to the time the version last changed, as opposed
to when the release number changed, to signify major differences. This is
extracted from the tag data and analysing the spec at that point in time.

Combined with the latest ypkg change to emit `SOURCE_DATE_EPOCH`, this
should remove much of the jitter in between package rebuilds and restore
sane deltas in no time at all.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>